Cherry-pick: [Android WebView] Fix bugs with initial-scale between 0 and 1
Cherry-pick of chromium crrev.com/24769002

Bug: 10742856

Original description:

The fix in https://chromiumcodereview.appspot.com/23754026 missed the
case of initial-scale, so fractional values were getting clamped to
zero incorrectly. Secondly, initial-scale values should not be clamped
in WebView.

BUG=298962, 282130, 295886
Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=158476

+++ Source/core/dom/Document.cpp
@@ -2828,7 +2828,7 @@ void Document::processViewport(const String& features, ViewportArguments::Type o
     processArguments(features, (void*)&m_viewportArguments, &setViewportFeature);
 
     if (page() && page()->settings()->viewportMetaZeroValuesQuirk() && m_viewportArguments.type == ViewportArguments::ViewportMeta
-        && m_viewportArguments.width == ViewportArguments::ValueDeviceWidth && !static_cast<int>(m_viewportArguments.zoom))
+        && m_viewportArguments.width == ViewportArguments::ValueDeviceWidth && !m_viewportArguments.zoom)
         m_viewportArguments.zoom = 1.0;
 
     updateViewportArguments();
